+++ mod/Location.cs
@@ -184,8 +184,8 @@ public static class LocationNames
     public static Dictionary<Location, SignalName> locationToSignal = signalToLocation.ToDictionary(stl => stl.Value, stl => stl.Key);
 
     // leave these as null until we load the ids, so any attempt to work with ids before that will fail loudly
-    public static Dictionary<int, Location> archipelagoIdToLocation = null;
-    public static Dictionary<Location, int> locationToArchipelagoId = null;
+    public static Dictionary<long, Location> archipelagoIdToLocation = null;
+    public static Dictionary<Location, long> locationToArchipelagoId = null;
 
     public static void LoadArchipelagoIds(string locationsFilepath)
     {
@@ -197,7 +197,7 @@ public static class LocationNames
             // Skip event locations, since they intentionally don't have ids
             if (locationData["address"].Type == JTokenType.Null) continue;
 
-            var archipelagoId = (int)locationData["address"];
+            var archipelagoId = (long)locationData["address"];
             var location = NameToLocation((string)locationData["name"]);
             archipelagoIdToLocation.Add(archipelagoId, location);
             locationToArchipelagoId.Add(location, archipelagoId);
